# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CST Portal Internal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reset-token:
    post:
      operationId: generateResetToken
      tags: [users]
      summary: Generate a password-reset token for a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetTokenResponse"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    User:
      type: object
      required: [id, name, email, role, dept, initials, color, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, sector_manager, employee]
        dept:
          type: string
        initials:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [ativo, inativo]
    UserInput:
      type: object
      required: [name, email, password, role, dept]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, sector_manager, employee]
        dept:
          type: string
        color:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, sector_manager, employee]
        dept:
          type: string
        status:
          type: string
          enum: [ativo, inativo]
        color:
          type: string
...
    ResetTokenResponse:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
        expiresAt:
          type: string
...