# lib/api-zod/src/generated/types/resetTokenResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CST Portal Internal API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetTokenResponse {
  token: string;
  expiresAt: string;
}